Wrapping and layout draft
This is my Mockup draft for what i imagine the Profile Generator web
app can look like:

- Index.html : Introduction to the problematic with centralized
profiles and explanation of what hosting your own profile can achieve.
If the user already has a profile, he/she can specify the url where is
hosted and it will be parsed for edit.

- creator.html: The form to create or edit profiles

- profile.html: The result once the profile has been created/edited,
available in several different formats. The tool would also give users
the possibility to NOT host their profiles themselves but leave them in
the platform.